requirements.txt
websocket-client==0.44.0
